Handling escaped characters in URLs
A bug was found, where escaped characters (e.g %e2%80%99) were not matched
properly by the redirected. This patch should ensure that any URLs, both
those not found and set up as old uri in the system, are matched (the
implemented fix is that all urls are decoded before they are matched).